+++ src/api/qualicharge/models/dynamic.py
@@ -6,11 +6,13 @@ from typing import Annotated, Optional
 from zoneinfo import ZoneInfo
 
 from pydantic import AfterValidator, PositiveFloat, model_validator
-from pydantic.types import PastDatetime
+from pydantic.types import AwareDatetime, PastDatetime
 from sqlmodel import Field, SQLModel
 from typing_extensions import Self
 
 from ..conf import settings
+
+AwarePastDatetime = Annotated[AwareDatetime, PastDatetime]
 
 
 def is_aware(value: datetime):
@@ -85,7 +87,7 @@ class StatusCreate(StatusAPIBase):
     """Point of charge status create."""
 
     horodatage: Annotated[
-        PastDatetime,
+        AwarePastDatetime,
         AfterValidator(
             lambda v: not_older_than(v, timedelta(seconds=settings.API_MAX_STATUS_AGE))
         ),
@@ -121,12 +123,9 @@ class SessionCreate(SessionBase):
         },
     )
     start: Annotated[
-        PastDatetime,
+        AwarePastDatetime,
         AfterValidator(
             lambda v: not_older_than(v, timedelta(seconds=settings.API_MAX_SESSION_AGE))
         ),
     ]
-    end: Annotated[
-        PastDatetime,
-        AfterValidator(lambda v: v if is_aware(v) else set_default_tz(v)),
-    ]
+    end: AwarePastDatetime